Fix release webhooks
Fix webhooks used for release announcements, which had a bug caused by
not properly accounting for python's variable scoping/capture logic.

The on_commit hook ended up referring to the same object for each
webhook call, which meant only one of the webhooks ever triggered if
package had multiple webhooks defined, and that webhook would trigger
multiple times.
