+++ django/thunderstore/repository/tests/test_utils.py
@@ -1,7 +1,5 @@
 import pytest
-from django.db import transaction
 
-from thunderstore.core.utils import on_commit_or_immediate
 from thunderstore.repository.package_reference import PackageReference
 from thunderstore.repository.utils import does_contain_package, has_duplicate_packages
 
@@ -93,19 +91,3 @@ def test_utils_does_contain_package(collection, reference, expected):
 def test_utils_has_duplicate_packages(collection, expected):
     collection = [PackageReference.parse(x) for x in collection]
     assert has_duplicate_packages(collection) == expected
-
-
-@pytest.mark.django_db(transaction=True)
-def test_utils_on_commit_or_immediate():
-    calls = []
-
-    def test_fn():
-        calls.append(1)
-
-    with transaction.atomic():
-        assert len(calls) == 0
-        on_commit_or_immediate(lambda: test_fn())
-        assert len(calls) == 0
-    assert len(calls) == 1
-    on_commit_or_immediate(lambda: test_fn())
-    assert len(calls) == 2
